refactor: rename jwt::log module to jwt::log_entry
